% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_handlers.R
\name{convert_node_data}
\alias{convert_node_data}
\title{Function for converting the data.entries component of a node on the Sankey visualization to an
R data.frame}
\usage{
convert_node_data(x, session, inputname)
}
\arguments{
\item{x}{The data obtained from the JavaScript Sankey visualization for an alt-clicked node.
This should already have been transformed with \code{jsonlite::fromJSON} and be a list of
lists.}

\item{session, inputname}{shiny session and inputname for the element. Not used.}
}
\value{
A data.frame with as many rows as there are entries in the list \code{x}. The names in
each sub-entry of \code{x} define the columns of this data.frame.
}
\description{
Function for converting the data.entries component of a node on the Sankey visualization to an
R data.frame
}
