% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_multiply.R
\name{dummy_multiply}
\alias{dummy_multiply}
\title{Multiply two numbers together}
\usage{
dummy_multiply(a, b)
}
\arguments{
\item{a, b}{Two numbers to multiply}
}
\value{
The product of `a` and `b`
}
\description{
This is a dummy function and should be removed
}
