% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_example.R
\name{run_shiny_example}
\alias{run_shiny_example}
\title{Run one of the example shiny apps associated with this package}
\usage{
run_shiny_example(app_name = "alt_click")
}
\arguments{
\item{app_name}{The name of the app to launch. See the "shiny-examples" directory of the
installed package.}
}
\description{
Run one of the example shiny apps associated with this package
}
