% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{sankey}
\alias{sankey}
\title{Create a Sankey plot}
\usage{
sankey(message, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{message}{Placeholder.}

\item{width, height}{The initial size of the visualization}

\item{elementId}{Identifier for the HTML element into which the visualization will be added.}
}
\value{
A html widget containing a Sankey diagram of the data
}
\description{
Create a HTML sankey plot for displaying admissions and retention data
}
