% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{sankey}
\alias{sankey}
\title{Create a Sankey plot}
\usage{
sankey(
  data,
  steps,
  color = NULL,
  hover_color = NULL,
  color_overrides = NULL,
  node_template = NULL,
  link_template = NULL,
  alt_click_handler = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Data-frame containing one row per observation (student).}

\item{steps}{Character vector or list containing a subset of the column names in \code{data}.
These correspond to the stages of the sankey chart. Where a list is provided, elements of the
list can contain vectors which specify how a step will be further subdivided (drilled-down
into) when that step is clicked on.}

\item{color}{Character representing a single color. This color is used to color all nodes
that don't have an associated color override. Links emanating from these nodes will be a
translucent version of this color.}

\item{hover_color}{Character representing a single color. This color is used to highlight a
node on hover. It is also used---in a translucent form---to highlight a link on hover.}

\item{color_overrides}{If supplied, this should be a list of lists. Each of the inner lists
specifies some form of color override. To have any effect it must have a "color" entry and at
least one of a "name" entry and a "group" entry. The "name" entry corresponds to the name or
label of a node while the "group" entry corresponds to the name associated with the column in
the Sankey. This means the groups are derived from the keys in the supplied data while the
names are the values associated with these keys. An override assigned to a node through a
"name" entry trumps an override assigned through a "group" entry. An override assigned through
a "name" and a "group" entry in the same object will trump both of these.}

\item{node_template, link_template}{Template strings defining the text presented the popups
that are shown when hovering over a node or a link. These use the
\href{https://handlebarsjs.com/}{Handlebars HTML templating language}. If supplied, these should be
a single character string. Alongside regular HTML constructs, the templates can reference the
following properties.

For \code{node_template}:
\itemize{
\item \code{ {{name}} } the name of the node.
\item \code{ {{count}} } the count of the node.
\item \code{ {{totalCount}} } the total count for the whole Sankey diagram.
}

For \code{link_template}:
\itemize{
\item \code{ {{sourceName}} } the name of the source node of the link.
\item \code{ {{targetName}} } the name of the target node of the link.
\item \code{ {{count}} } the count of the link.
\item \code{ {{totalCount}} } the total count for the whole Sankey diagram.
\item \code{ {{percentageOfSourceCount}} } the (rounded) percentage value of all counts from the
source node that go through this link.
\item \code{ {{percentageOfTargetCount}} } the (rounded) percentage value of all counts to the
target node that go through this link.
}}

\item{alt_click_handler}{A JavaScript function to be called when the user alt-clicks on a
node in the Sankey chart. Construct this with \code{htmlwidgets::JS()}. The function should
have parameters "event" and "data".}

\item{width, height}{The initial size of the visualization}

\item{elementId}{Identifier for the HTML element into which the visualization will be added.}
}
\value{
A html widget containing a Sankey diagram of the data
}
\description{
Create a HTML sankey plot for displaying admissions and retention data
}
\examples{
data(admissions)
sankey(
  admissions,
  steps = c("student_type", "gpa", "outcome")
)

# steps can be a list containing vectors.
sankey(
  admissions,
  steps = list(c("student_type", "college"), "gpa", "outcome")
)

}
